Refactor primarily around config POCOs
