+++ src/WorkflowCoordinator/WorkflowCoordinator/Config/GeneralConfig.cs
@@ -4,6 +4,8 @@ namespace WorkflowCoordinator.Config
 {
     public class GeneralConfig
     {
+        public Uri DataAccessLocalhostBaseUri { get; set; }
+
         public Uri AzureDbTokenUrl { get; set; }
     }
 }
