+++ src/WorkflowCoordinator/WorkflowCoordinator/Config/StartupConfig.cs
@@ -5,7 +5,7 @@ namespace WorkflowCoordinator.Config
     public class StartupConfig
     {
         public string EndpointName { get; set; }
-        public Uri DataAccessWebServiceLocalhostBaseUri { get; set; }
+        public Uri DataAccessLocalhostBaseUri { get; set; }
 
     }
 }
