+++ src/WorkflowCoordinator/WorkflowCoordinator/HttpClients/IDataServiceApiClient.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using WorkflowCoordinator.Models;
+
+namespace WorkflowCoordinator.HttpClients
+{
+    public interface IDataServiceApiClient
+    {
+        Task<IEnumerable<AssessmentModel>> GetAssessments(string callerCode);
+    }
+}
