+++ src/WorkflowCoordinator/WorkflowCoordinator/Models/AssessmentModel.cs
@@ -1,0 +1,9 @@
+﻿namespace WorkflowCoordinator.Models
+{
+    public class AssessmentModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string SourceName { get; set; }
+    }
+}
